+++ CHANGELOG.md
@@ -7,6 +7,16 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.1.2] - 2023-02-12
+
+### Added
+
+- files resideInAPath() assert/filter/dependency check
+
+### Fixed
+
+- Files dependency check incorrect imports resolve
+
 ## [0.1.1] - 2023-02-11
 
 ### Added